object-based-system-raft/ui-svc/requirements.txt
flask
grpcio
grpcio-tools
protobuf
